requirements.txt
Flask>=2.3.0
flask-session>=0.5.0
supabase>=2.0.0
python-dotenv>=1.0.0

